Add a simple example to create an S3 bucket
Simple example that creates a private AWS S3 bucket intended to be
bootstrapped via terraform012 and then updated.

terraform-aws-modules/s3-bucket/aws is pinned to 2.9.0 because it was
the last version to work with Terraform 0.12.x.

This is intended to be used via moto_server provided by net/py-moto
package.